Replaces 'writer' with generic 'visitor'.
This commit removes the hand-written 'tapl.arith.writer' module and replaces it
with a combination of a generic 'tapl.visit' module along with
'tapl.arith.formatters.TextFormatter' which is able to leverage the visiting
routines to produce output.

Additional formatters (and visitors) may be introduced without requiring
duplication of the visiting logic.  This requires the declaration of which
properties of the terms are themselves terms; this is done through the addition
of the 'subterms' class property on each term.