tapl.lrparser: 'producer' functions
Instead of constructing an explicit concrete syntax tree which must then be
transformed to an abstract syntax tree, a 'producer' function is invoked.  The
'terms' module is expected to contain a list of producer functions which
correspond to each reduction rule in the grammar.  Any required transformations
may be done by these producer functions.