tapl: adds 'toolchain'
This commit performs a large overhaul of the internal structure of 'tapl'.

First, each implementation supplies a 'Toolchain' class, which provides some
static data and methods that describe its language.  This includes a list of
token regexes, an LR(1) table, a list of what to produce for each reduction
rule, a semantic analyzer and an evaluator.  This is preferred over having
separate modules in a package which each provide this data.

Second, an LR(1) parser generator is included in 'tools/make_table.py', this
has been leveraged to generate the LR(1) tables included in each 'toolchain'.
An appropriately formatted grammar is also included for each implementation.

Third, each implementation may provide a 'formatters' package, each containing
a module providing a 'Formatter' class that can be used to emit text given a
semantic tree.

Finally, 'driver.py' disappears since its primary purpose was to mechanically
provide what 'Toolchain' now provides.