tools: refactor LR table tools
This commit introduces several new 'tool' scripts, and isolates their
functionality:

1. 'build_table' replaces 'make_table' and is responsible only for the actual
table building.

2. 'write_table' uses functions from 'build_table' to produce a Python 'dict'
literal that may be used in a 'toolchain' for an interpreter.  This replaces
the original functionality of 'make_table' when run without the '-d' switch.

3. 'document_table' operates much like 'write_table' from (2) but implements
instead the original behaviour of 'make_table' when run with the '-d' switch.

4. 'format_table' is renamed to 'format_literal' to better reflect what it
does, since it is agnostic to LR tables.