relexer: consume tuple not dict of tokens
The set of valid tokens are now specified as a tuple instead of a dict.  Tuples
provide a specification of order of matching, allowing more general regular
expressions to be placed lower in the list from more specific expressions.
This allows, for example, identifiers to be specified as a sequence of
characters after keywords as exact sequences without any ambiguities.