toolchains: remove unnecessary 'coding' directive
The 'UTF-8' 'coding' directives were formerly required to facilitate inline
documentation of the LR table generation.  Since this is now automated by means
of the LR table tools, they can be obtained separately as required.  As a
result, the inline documentation has been removed, and the 'coding' directives
are no longer required.