table: move table literals to separate files
Moving the table literals out of the 'toolchain' modules allows for faster
iterating on both, the grammar rules, and the LR table builder, since it does
not require modification and formatting of the table within the 'toolchain'
module.

A helper tool, 'write_table_literal' is added to facilitate production of an
appropriate 'table' module for a given 'Toolchain' to reference.